+++ tensorcircuit/__init__.py
@@ -18,7 +18,7 @@ from .mpscircuit import MPSCircuit
 from .densitymatrix import DMCircuit
 from .densitymatrix2 import DMCircuit2
 from .gates import num_to_tensor, array_to_tensor
-from .vis import qir2tex
+from .vis import qir2tex, render_pdf
 from . import interfaces
 from . import templates
 from . import quantum