+++ tensorcircuit/__init__.py
@@ -15,8 +15,10 @@ from .cons import (
 from . import gates
 from .circuit import Circuit, expectation
 from .mpscircuit import MPSCircuit
-from .densitymatrix import DMCircuit
+from .densitymatrix import DMCircuit as DMCircuit_reference
 from .densitymatrix2 import DMCircuit2
+
+DMCircuit = DMCircuit2  # compatibility issue to still expose DMCircuit2
 from .gates import num_to_tensor, array_to_tensor
 from .vis import qir2tex, render_pdf
 from . import interfaces