+++ tensorcircuit/backends/abstract_backend.py
@@ -1223,3 +1223,5 @@ def _more_methods_for_backend(tnbackend: Any) -> None:
 
 
 _more_methods_for_backend(tnbackend)
+
+# TODO(@refraction-ray): implement __str__ and __repr__ for backends