+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.0.220118
+
 ### Added
 
 - add `hessian` method on backends